Extract separate class for applying TripUpdates to Timetables
