+++ application/src/test/java/org/opentripplanner/updater/trip/TimetableMapperTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.model;
+package org.opentripplanner.updater.trip;
 
 import static com.google.transit.realtime.GtfsRealtime.TripDescriptor.ScheduleRelationship.SCHEDULED;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -28,16 +28,16 @@ import org.junit.jupiter.api.Test;
 import org.opentripplanner.ConstantsForTests;
 import org.opentripplanner.TestOtpModel;
 import org.opentripplanner._support.time.ZoneIds;
+import org.opentripplanner.model.Timetable;
+import org.opentripplanner.model.TripTimesPatch;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
 import org.opentripplanner.transit.model.framework.Result;
 import org.opentripplanner.transit.model.network.TripPattern;
 import org.opentripplanner.transit.model.timetable.RealTimeState;
 import org.opentripplanner.transit.service.TimetableRepository;
 import org.opentripplanner.updater.spi.UpdateError;
-import org.opentripplanner.updater.trip.BackwardsDelayPropagationType;
-import org.opentripplanner.updater.trip.TripUpdateBuilder;
 
-public class TimetableTest {
+public class TimetableMapperTest {
 
   private static final ZoneId TIME_ZONE = ZoneIds.NEW_YORK;
   private static final LocalDate SERVICE_DATE = LocalDate.of(2009, 8, 7);
@@ -66,12 +66,6 @@ public class TimetableTest {
   }
 
   @Test
-  public void getGetTripTimes() {
-    var tt = timetable.getTripTimes(new FeedScopedId(feedId, TRIP_ID));
-    assertNotNull(tt);
-  }
-
-  @Test
   public void tripNotFoundInPattern() {
     // non-existing trip
     var tripDescriptorBuilder = tripDescriptorBuilder("b");
@@ -86,7 +80,7 @@ public class TimetableTest {
     stopTimeUpdateBuilder.setScheduleRelationship(StopTimeUpdate.ScheduleRelationship.NO_DATA);
     var tripUpdate = tripUpdateBuilder.build();
 
-    Result<TripTimesPatch, UpdateError> result = Timetable.createUpdatedTripTimesFromGTFSRT(
+    Result<TripTimesPatch, UpdateError> result = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -115,7 +109,7 @@ public class TimetableTest {
     stopTimeUpdateBuilder.setStopSequence(0);
     stopTimeUpdateBuilder.setScheduleRelationship(StopTimeUpdate.ScheduleRelationship.SKIPPED);
     tripUpdate = tripUpdateBuilder.build();
-    var result = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var result = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -145,7 +139,7 @@ public class TimetableTest {
       LocalDateTime.of(2009, Month.AUGUST, 7, 0, 10, 0, 0).atZone(TIME_ZONE).toEpochSecond()
     );
     var tripUpdate = tripUpdateBuilder.build();
-    var result = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var result = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -180,7 +174,7 @@ public class TimetableTest {
     );
     var tripUpdate = tripUpdateBuilder.build();
     assertEquals(20 * 60, timetable.getTripTimes(trip_1_1_index).getArrivalTime(2));
-    var result = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var result = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -208,7 +202,7 @@ public class TimetableTest {
     stopTimeEventBuilder.setDelay(0);
     tripUpdate = tripUpdateBuilder.build();
     result =
-      Timetable.createUpdatedTripTimesFromGTFSRT(
+      TimetableMapper.createUpdatedTripTimesFromGTFSRT(
         timetable,
         tripUpdate,
         TIME_ZONE,
@@ -238,7 +232,7 @@ public class TimetableTest {
     tripUpdate = tripUpdateBuilder.build();
 
     result =
-      Timetable.createUpdatedTripTimesFromGTFSRT(
+      TimetableMapper.createUpdatedTripTimesFromGTFSRT(
         timetable,
         tripUpdate,
         TIME_ZONE,
@@ -266,7 +260,7 @@ public class TimetableTest {
     stopTimeEventBuilder.setDelay(120);
     tripUpdate = tripUpdateBuilder.build();
     result =
-      Timetable.createUpdatedTripTimesFromGTFSRT(
+      TimetableMapper.createUpdatedTripTimesFromGTFSRT(
         timetable,
         tripUpdate,
         TIME_ZONE,
@@ -293,7 +287,7 @@ public class TimetableTest {
     stopTimeEventBuilder.setDelay(120);
     tripUpdate = tripUpdateBuilder.build();
     result =
-      Timetable.createUpdatedTripTimesFromGTFSRT(
+      TimetableMapper.createUpdatedTripTimesFromGTFSRT(
         timetable,
         tripUpdate,
         TIME_ZONE,
@@ -328,7 +322,7 @@ public class TimetableTest {
     stopTimeEventBuilder.setDelay(-1);
     var tripUpdate = tripUpdateBuilder.build();
 
-    var result = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var result = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -354,7 +348,7 @@ public class TimetableTest {
     stopTimeUpdateBuilder.setStopSequence(3);
     stopTimeUpdateBuilder.setScheduleRelationship(StopTimeUpdate.ScheduleRelationship.NO_DATA);
     TripUpdate tripUpdate = tripUpdateBuilder.build();
-    var result = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var result = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -398,7 +392,7 @@ public class TimetableTest {
     stopTimeEventBuilder = stopTimeUpdateBuilder.getArrivalBuilder();
     stopTimeEventBuilder.setDelay(15);
     TripUpdate tripUpdate = tripUpdateBuilder.build();
-    var result = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var result = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -437,7 +431,7 @@ public class TimetableTest {
     stopTimeEventBuilder.setDelay(15);
     TripUpdate tripUpdate = tripUpdateBuilder.build();
 
-    var result = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var result = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -470,7 +464,7 @@ public class TimetableTest {
     StopTimeEvent.Builder stopTimeEventBuilder = stopTimeUpdateBuilder.getArrivalBuilder();
     stopTimeEventBuilder.setDelay(-100);
     TripUpdate tripUpdate = tripUpdateBuilder.build();
-    var patch = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var patch = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -510,7 +504,7 @@ public class TimetableTest {
     StopTimeEvent.Builder stopTimeEventBuilder = stopTimeUpdateBuilder.getArrivalBuilder();
     stopTimeEventBuilder.setDelay(-700);
     TripUpdate tripUpdate = tripUpdateBuilder.build();
-    var patch = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var patch = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -555,7 +549,7 @@ public class TimetableTest {
     stopTimeEventBuilder = stopTimeUpdateBuilder.getArrivalBuilder();
     stopTimeEventBuilder.setDelay(15);
     TripUpdate tripUpdate = tripUpdateBuilder.build();
-    var result = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var result = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -584,7 +578,7 @@ public class TimetableTest {
     stopTimeEventBuilder.setDelay(-700);
     TripUpdate tripUpdate = tripUpdateBuilder.build();
 
-    var result = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var result = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -634,7 +628,7 @@ public class TimetableTest {
 
     TripUpdate tripUpdate = tripUpdateBuilder.build();
 
-    var patch = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var patch = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -680,7 +674,7 @@ public class TimetableTest {
 
     TripUpdate tripUpdate = tripUpdateBuilder.build();
 
-    var patch = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var patch = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -725,7 +719,7 @@ public class TimetableTest {
 
     TripUpdate tripUpdate = tripUpdateBuilder.build();
 
-    var patch = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var patch = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -760,7 +754,7 @@ public class TimetableTest {
 
     TripUpdate tripUpdate = tripUpdateBuilder.build();
 
-    var patch = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var patch = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -805,7 +799,7 @@ public class TimetableTest {
 
     TripUpdate tripUpdate = tripUpdateBuilder.build();
 
-    var patch = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var patch = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       timetable,
       tripUpdate,
       TIME_ZONE,
@@ -858,7 +852,7 @@ public class TimetableTest {
     var scheduledTimetable = patternIndex
       .get(new FeedScopedId(feedId, TRIP_ID_WITH_MORE_STOPS))
       .getScheduledTimetable();
-    var patch = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var patch = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       scheduledTimetable,
       tripUpdate,
       TIME_ZONE,
@@ -922,7 +916,7 @@ public class TimetableTest {
       .get(new FeedScopedId(feedId, TRIP_ID_WITH_MORE_STOPS))
       .getScheduledTimetable();
 
-    var patch = Timetable.createUpdatedTripTimesFromGTFSRT(
+    var patch = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
       scheduledTimetable,
       tripUpdate,
       TIME_ZONE,
@@ -975,7 +969,7 @@ public class TimetableTest {
     }
 
     private static void testInvalidStopTime(
-      BiConsumer<TripUpdate.StopTimeUpdate.Builder, StopTimeEvent> setEmptyEvent,
+      BiConsumer<StopTimeUpdate.Builder, StopTimeEvent> setEmptyEvent,
       UpdateError.UpdateErrorType expectedError
     ) {
       var builder = new TripUpdateBuilder(TRIP_ID, SERVICE_DATE, SCHEDULED, TIME_ZONE);
@@ -983,7 +977,7 @@ public class TimetableTest {
       builder.addRawStopTime(emptyStopTime(2, setEmptyEvent));
       TripUpdate tripUpdate = builder.build();
 
-      var result = Timetable.createUpdatedTripTimesFromGTFSRT(
+      var result = TimetableMapper.createUpdatedTripTimesFromGTFSRT(
         timetable,
         tripUpdate,
         TIME_ZONE,
@@ -1001,7 +995,7 @@ public class TimetableTest {
 
     private static StopTimeUpdate emptyStopTime(
       int sequence,
-      BiConsumer<TripUpdate.StopTimeUpdate.Builder, StopTimeEvent> setEmptyEvent
+      BiConsumer<StopTimeUpdate.Builder, StopTimeEvent> setEmptyEvent
     ) {
       var emptyEvent = StopTimeEvent.newBuilder().build();
       var stopTime = StopTimeUpdate.newBuilder();
